refactor: move speed factor slider into toolhead controls
- remove all PrintsettingsPanel related code, that panel does not exist anymore
